+++ tests/tools/test_pip_tools.py
@@ -4,12 +4,18 @@
 
 from __future__ import annotations
 
+import subprocess
+import sys
+from contextlib import redirect_stderr, redirect_stdout
+from io import StringIO
 from pathlib import Path
 from tempfile import TemporaryDirectory
-from unittest import TestCase
+from typing import Any
+from unittest import mock, TestCase
 
 from license_tools.tools import pip_tools
 from license_tools.utils import archive_utils
+from license_tools.utils.path_utils import get_files_from_directory
 from tests import get_from_url
 from tests.data import JWCRYPTO__1_5_4__TAR_GZ, PYPDF__3_17_4__WHEEL
 
@@ -98,3 +104,131 @@ License classifier: BSD License
            Summary: Implementation of JOSE Web standards
 License classifier:
 """[1:-1], result)
+
+
+class DownloadPackageTestCase(TestCase):
+    def test_valid_package_name(self) -> None:
+        stderr = StringIO()
+
+        with redirect_stderr(stderr), TemporaryDirectory() as directory:
+            pip_tools.download_package(
+                package_definition="typing_extensions==4.8.0",
+                index_url="https://pypi.org/simple",
+                download_directory=directory
+            )
+            actual = [x[1] for x in get_files_from_directory(directory)]
+            self.assertEqual(
+                ["typing_extensions-4.8.0-py3-none-any.whl"],
+                actual
+            )
+
+        self.assertEqual("", stderr.getvalue())
+
+    def test_prefer_source_distribution(self) -> None:
+        stderr = StringIO()
+
+        with redirect_stderr(stderr), TemporaryDirectory() as directory:
+            pip_tools.download_package(
+                package_definition="typing_extensions==4.8.0",
+                index_url="https://pypi.org/simple",
+                download_directory=directory,
+                prefer_sdist=True
+            )
+            actual = [x[1] for x in get_files_from_directory(directory)]
+            self.assertEqual(
+                ["typing_extensions-4.8.0.tar.gz"],
+                actual
+            )
+
+        self.assertEqual("", stderr.getvalue())
+
+    def test_invalid_package_name(self) -> None:
+        stdout, stderr = StringIO(), StringIO()
+
+        with redirect_stdout(stdout), redirect_stderr(stderr), TemporaryDirectory() as directory:
+            with self.assertRaises(subprocess.CalledProcessError):
+                pip_tools.download_package(
+                    package_definition="typing_extensions==1234567890",
+                    index_url="https://pypi.org/simple",
+                    download_directory=directory
+                )
+            actual = [x[1] for x in get_files_from_directory(directory)]
+            self.assertEqual(
+                [],
+                actual
+            )
+
+        stderr_string = stderr.getvalue()
+        self.assertEqual("", stdout.getvalue())
+        self.assertIn(
+            "ERROR: Could not find a version that satisfies the requirement typing_extensions==1234567890 (from versions: ",
+            stderr_string,
+        )
+        self.assertIn(
+            "\nERROR: No matching distribution found for typing_extensions==1234567890\n",
+            stderr_string,
+        )
+
+    def test_index_url_handling(self) -> None:
+        directories = []
+
+        def run(
+            command: list[str | Path], *args: Any, **kwargs: Any
+        ) -> "subprocess.CompletedProcess[bytes]":
+            _directory = command[command.index("--dest") + 1]
+            directories.append(_directory)
+            Path(_directory).joinpath("dummy.py").touch()
+            return subprocess.CompletedProcess(args=args, returncode=0, stdout=b"")
+
+        with mock.patch("subprocess.run", side_effect=run) as subprocess_mock:
+            with TemporaryDirectory() as directory:
+                pip_tools.download_package(
+                    package_definition="testing", download_directory=directory
+                )
+            with TemporaryDirectory() as directory:
+                pip_tools.download_package(
+                    package_definition="testing",
+                    index_url="DUMMY",
+                    download_directory=directory,
+                )
+
+            subprocess_mock.assert_has_calls(
+                [
+                    mock.call(
+                        [
+                            sys.executable,
+                            "-m",
+                            "pip",
+                            "download",
+                            "--no-deps",
+                            "testing",
+                            "--dest",
+                            directories[0],
+                        ],
+                        capture_output=True,
+                        text=True,
+                        check=True,
+                    ),
+                    mock.call(
+                        [
+                            sys.executable,
+                            "-m",
+                            "pip",
+                            "download",
+                            "--no-deps",
+                            "testing",
+                            "--dest",
+                            directories[1],
+                            "--index-url",
+                            "DUMMY",
+                        ],
+                        capture_output=True,
+                        text=True,
+                        check=True,
+                    ),
+                ],
+                any_order=False,
+            )
+            self.assertEqual(
+                2, subprocess_mock.call_count, subprocess_mock.call_args_list
+            )