+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Development version
 
+* Prefer logging over printing.
+* Move download functionality to dedicated methods.
+* Make `Cargo.toml` metadata parsing optional.
+
 # Version 0.10.0 - 2024-04-19
 
 * Add handling for Rust crates